Move diffsky/phot_utils.py into diffsky/utils and fix imports
