+++ diffsky/__init__.py
@@ -5,4 +5,4 @@
 from ._version import __version__
 from .mass_functions.mc_diffmah_tpeak import mc_subhalos
 from .soft_histograms import diffndhist, diffndhist_lomem, signdhist, signdhist_lomem
-from .utils import tw_utils
+from .utils import phot_utils, tw_utils
